+++ app/recipes/recipes.$recipeId_.edit.route.tsx
@@ -11,9 +11,9 @@ import { useEffect, useRef, useState } from "react";
 import { recipeAction } from "~/recipes/recipe-actions";
 import { loadSingleRecipe } from "~/recipes/recipe-loader";
 import { FormTextAreaInput, FormTextInput } from "~/components/forms";
-import { SubmissionStyles } from "~/components/recipes";
-import { useIngredientsForm } from "~/components/recipes/use-ingredients-form";
-import { useKeyboardSubmit } from "~/components/use-keyboard";
+import { SubmissionStyles } from "~/recipes/recipes";
+import { useIngredientsForm } from "~/recipes/recipes/use-ingredients-form";
+import { useKeyboardSubmit } from "~/hooks/use-keyboard";
 import VisuallyHidden from "~/components/visually-hidden";
 import { getDefaultRecipeValues } from "~/utils";
 import { isValidString } from "~/utils/strings";
