+++ app/users/api.restricted.route.tsx
